+++ src/formatter/formatter.c
@@ -54,7 +54,7 @@ PUREFUNC text_opts_t choose_text_options(ast_list_t *chunks) {
 }
 
 static bool starts_with_id(Text_t text) {
-    List_t codepoints = Text$utf32_codepoints(Text$slice(text, I_small(1), I_small(1)));
+    List_t codepoints = Text$utf32(Text$slice(text, I_small(1), I_small(1)));
     return uc_is_property_xid_continue(*(ucs4_t *)codepoints.data);
 }