+++ CHANGES.md
@@ -9,6 +9,10 @@
 - List indexing now gives an optional value
 - Added a `--format` flag to the `tomo` binary that autoformats your code
   (currently unstable, do not rely on it just yet).
+- Standardized text methods for Unicode encodings:
+  - `Text.from_utf8()`/`Text.utf8()`
+  - `Text.from_utf16()`/`Text.utf16()`
+  - `Text.from_utf32()`/`Text.utf32()`
 - Fixed bugs:
   - `Int.parse()` had a memory bug.